chore/#12 BoardItem 컴포넌트 분리
