+++ src/components/NoteList.jsx
@@ -5,6 +5,7 @@ import { ReactComponent as IPlus } from '../assets/iPlus.svg';
 import { ReactComponent as IOption } from '../assets/iOption.svg';
 import { Link } from 'react-router-dom';
 import DropBox from '../components/common/DropBox';
+import BoardItem from './common/BoardItem';
 
 const noteList = [
   { title: '제목1', date: '2022년 11월 14일 (월)', todo: '할 일 2개' },
@@ -70,18 +71,6 @@ function NoteList() {
         ))}
       </StyledNoteList>
     </StyledRoot>
-  );
-}
-
-function BoardItem() {
-  return (
-    <StyledBoardItemWrapper>
-      {Array.from({ length: 3 }, (_v, i) => i).map((idx) => (
-        <StyledBoardItem key={idx} idx={idx}>
-          <div></div>
-        </StyledBoardItem>
-      ))}
-    </StyledBoardItemWrapper>
   );
 }
 
@@ -155,31 +144,4 @@ const StyledNoteInfo = styled.div`
   font-size: 1.4rem;
   line-height: 1.7rem;
   color: ${pinterestColors.gray400};
-`;
-
-const StyledBoardItemWrapper = styled.article`
-  display: grid;
-  grid-template-columns: 16.8rem;
-  grid-template-rows: 9.5rem 9.3rem;
-  border-radius: 1rem;
-  overflow: hidden;
-  margin-top: 2.2rem;
-`;
-
-const StyledBoardItem = styled.div`
-  height: 18.8rem;
-  background-color: #d9d9d9;
-
-  ${({ idx }) =>
-    idx === 1 &&
-    css`
-      grid-column: 2/3;
-      background-color: #a9a9a9;
-    `}
-  ${({ idx }) =>
-    idx === 2 &&
-    css`
-      grid-column: 2/3;
-      background-color: #c0c0c0;
-    `}
 `;
